requirements.txt
Django==5.0.0
djangorestframework==3.14.0
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.6
gunicorn==21.2.0
python-dotenv==1.0.0
stripe==7.10.0
django-parler==2.3.0
django-redis==5.4.0
django-cors-headers==4.3.1
Pillow==10.1.0
mailchimp-marketing==3.0.80
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0
django-storages==1.14.2
whitenoise==6.6.0